fixed game section and score section
